requirements.txt
Flask==3.0.2
azure-ai-inference==1.0.0b9
python-dotenv==1.0.1
SQLAlchemy==2.0.28
pandas==2.2.1
gunicorn==21.2.0
azure-core==1.34.0
mysql-connector-python==8.3.0 